Split build command execution tests
